Local and Global scopes
A variable is only available from inside the region it is created. This is called scope.

A variable created inside a function belongs to the local scope of that function, and can only be used inside that function.

A variable created in the main body of the Python code is a global variable and belongs to the global scope.

If you need to create a global variable, but are stuck in the local scope, you can use the global keyword.

The global keyword makes the variable global.